feat: add currency card with code
