+++ src/components/ui/currencyCard/CurrencyCard.module.scss
@@ -1,13 +1,9 @@
 @import '@/styles/variables.scss';
 
 .currencyCard {
-  max-width: 520px;
+  max-width: 454px;
   width: 100%;
   display: flex;
-  padding: 34px 32px;
-  border: 1px solid $grey-400;
-  border-radius: 8px;
-  background: var(--secondary-color-200);
 }
 
 .icon {
@@ -31,11 +27,4 @@
   color: var(--text-color-200);
   font-size: 35px;
   line-height: 1.3;
-}
-
-.rate {
-  color: var(--text-color-300);
-  font-size: 32px;
-  font-weight: 300;
-  line-height: 1;
 }
