+++ src/components/ui/currencyCodeCard/CurrencyCodeCard.module.scss
@@ -1,0 +1,6 @@
+.code {
+  color: var(--text-color-300);
+  font-size: 16px;
+  font-weight: 500;
+  line-height: 1;
+}
