+++ src/components/ui/currencyRatedCard/CurrencyRatedCard.module.scss
@@ -1,0 +1,17 @@
+@import '@/styles/variables.scss';
+
+.currencyCard {
+  max-width: 520px;
+  width: 100%;
+  padding: 34px 32px;
+  border: 1px solid $grey-400;
+  border-radius: 8px;
+  background: var(--secondary-color-200);
+}
+
+.rate {
+  color: var(--text-color-300);
+  font-size: 32px;
+  font-weight: 300;
+  line-height: 1;
+}
